Feature/jwtdecode (#143)
* decode jwt

* split services

* split services and update calls

* fix typo

* decode jwt

* split services

* split services and update calls

* fix typo

* chore: cleanup some tests and references to fake jwt that do not decode

* remove services file

* remove ad_id decode and header as it is not needed by fsbid
