+++ talentmap_api/fsbid/tests/test_fsbid_bid_seasons.py
@@ -3,8 +3,6 @@ from model_mommy import mommy
 from unittest.mock import Mock, patch
 from rest_framework import status
 
-import talentmap_api.fsbid.services as services
-
 bs = {
     "bsn_id": "1",
     "bsn_descr_text": "Test Bid Season",
@@ -13,6 +11,7 @@ bs = {
     "bsn_panel_cutoff_date": "2020/04/15"
 }
 
+fake_jwt = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1bmlxdWVfbmFtZSI6IldBU0hEQ1xcVEVTVFVTRVIifQ.o5o4XZ3Z_vsqqC4a2tGcGEoYu3sSYxej4Y2GcCQVtyE"
 
 @pytest.fixture
 def test_bidder_fixture(authorized_user):
@@ -23,8 +22,8 @@ def test_bidder_fixture(authorized_user):
 @pytest.mark.django_db(transaction=True)
 @pytest.mark.usefixtures("test_bidder_fixture")
 def test_bid_seasons_actions(authorized_client, authorized_user):
-    with patch('talentmap_api.fsbid.services.requests.get') as mock_get:
+    with patch('talentmap_api.fsbid.services.bid_season.requests.get') as mock_get:
         mock_get.return_value = Mock(ok=True)
         mock_get.return_value.json.return_value = [bs]
-        response = authorized_client.get(f'/api/v1/fsbid/bid_seasons', HTTP_JWT='fake JWT')
+        response = authorized_client.get(f'/api/v1/fsbid/bid_seasons', HTTP_JWT=fake_jwt)
         assert response.json()[0]['id'] == [bs][0]['bsn_id']
