+++ talentmap_api/fsbid/tests/test_fsbid_bidlist.py
@@ -6,8 +6,6 @@ from unittest.mock import Mock, patch
 from rest_framework import status
 from django.utils import timezone
 
-import talentmap_api.fsbid.services as services
-
 bid = {
     "submittedDate": "2019/01/01",
     "statusCode": "A",
@@ -30,6 +28,7 @@ bid = {
     }
 }
 
+fake_jwt = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1bmlxdWVfbmFtZSI6IldBU0hEQ1xcVEVTVFVTRVIifQ.o5o4XZ3Z_vsqqC4a2tGcGEoYu3sSYxej4Y2GcCQVtyE"
 
 @pytest.fixture
 def test_bidder_fixture(authorized_user):
@@ -40,34 +39,34 @@ def test_bidder_fixture(authorized_user):
 @pytest.mark.django_db(transaction=True)
 @pytest.mark.usefixtures("test_bidder_fixture")
 def test_bidlist_actions(authorized_client, authorized_user):
-    with patch('talentmap_api.fsbid.services.requests.get') as mock_get:
+    with patch('talentmap_api.fsbid.services.bid.requests.get') as mock_get:
         mock_get.return_value = Mock(ok=True)
         mock_get.return_value.json.return_value = [bid]
-        response = authorized_client.get(f'/api/v1/fsbid/bidlist/', HTTP_JWT='fake JWT')
+        response = authorized_client.get(f'/api/v1/fsbid/bidlist/', HTTP_JWT=fake_jwt)
         assert response.json()[0]['emp_id'] == [bid][0]['employee']['perdet_seq_num']
 
 
 @pytest.mark.django_db(transaction=True)
 @pytest.mark.usefixtures("test_bidder_fixture")
 def test_bidlist_position_actions(authorized_client, authorized_user):
-    with patch('talentmap_api.fsbid.services.requests.get') as mock_get:
+    with patch('talentmap_api.fsbid.services.bid.requests.get') as mock_get:
         # returns 404 when no position is found
         mock_get.return_value = Mock(ok=True)
         mock_get.return_value.json.return_value = []
-        response = authorized_client.get(f'/api/v1/fsbid/bidlist/position/1/', HTTP_JWT='fake JWT')
+        response = authorized_client.get(f'/api/v1/fsbid/bidlist/position/1/', HTTP_JWT=fake_jwt)
         assert response.status_code == status.HTTP_404_NOT_FOUND
         # returns 204 when position is found
         mock_get.return_value = Mock(ok=True)
         mock_get.return_value.json.return_value = [bid]
-        response = authorized_client.get(f'/api/v1/fsbid/bidlist/position/1/', HTTP_JWT='fake JWT')
+        response = authorized_client.get(f'/api/v1/fsbid/bidlist/position/1/', HTTP_JWT=fake_jwt)
         assert response.status_code == status.HTTP_204_NO_CONTENT
 
-    with patch('talentmap_api.fsbid.services.requests.post') as mock_post:
+    with patch('talentmap_api.fsbid.services.bid.requests.post') as mock_post:
         mock_post.return_value = Mock(ok=True)
-        response = authorized_client.put(f'/api/v1/fsbid/bidlist/position/1/', HTTP_JWT='fake JWT')
+        response = authorized_client.put(f'/api/v1/fsbid/bidlist/position/1/', HTTP_JWT=fake_jwt)
         assert response.status_code == status.HTTP_204_NO_CONTENT
 
-    with patch('talentmap_api.fsbid.services.requests.delete') as mock_del:
+    with patch('talentmap_api.fsbid.services.bid.requests.delete') as mock_del:
         mock_del.return_value = Mock(ok=True)
-        response = authorized_client.delete(f'/api/v1/fsbid/bidlist/position/1/', HTTP_JWT='fake JWT')
+        response = authorized_client.delete(f'/api/v1/fsbid/bidlist/position/1/', HTTP_JWT=fake_jwt)
         assert response.status_code == status.HTTP_204_NO_CONTENT
